BlockListV20.py
Version 20 of the list is the last version I am adding tonight. I have noticed that the encoding on the file (on my end) was changed to UNIX (LF) compared to Windows (Carriage Return, Line Feed) this may have been done to another file, which explains the smaller sizes. Version 20 of the blocklist is another small update.